Remove ChildAttendanceContext
