+++ frontend/src/employee-mobile-frontend/child-attendance/AttendancePageWrapper.tsx
@@ -19,6 +19,7 @@ import {
   ChildrenResponse,
   GroupInfo
 } from 'lib-common/generated/api-types/attendance'
+import { useQueryResult } from 'lib-common/query'
 import useNonNullableParams from 'lib-common/useNonNullableParams'
 import { ContentArea } from 'lib-components/layout/Container'
 import colors from 'lib-customizations/common'
@@ -33,7 +34,8 @@ import { ChildAttendanceUIState, mapChildAttendanceUIState } from '../types'
 
 import AttendanceList from './AttendanceList'
 import ChildList from './ChildList'
-import { ChildAttendanceContext, ChildAttendanceStatuses } from './state'
+import { childrenQuery } from './queries'
+import { ChildAttendanceStatuses, useAttendanceStatuses } from './state'
 
 export default React.memo(function AttendancePageWrapper() {
   const { unitId, groupId, attendanceStatus } = useNonNullableParams<{
@@ -55,9 +57,8 @@ export default React.memo(function AttendancePageWrapper() {
     [groupId, unitInfoResponse]
   )
 
-  const { unitChildren, childAttendanceStatuses } = useContext(
-    ChildAttendanceContext
-  )
+  const unitChildren = useQueryResult(childrenQuery(unitId))
+  const childAttendanceStatuses = useAttendanceStatuses(unitId)
 
   const [showSearch, setShowSearch] = useState<boolean>(false)
