+++ frontend/src/employee-mobile-frontend/child-info/AttendanceChildPage.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components'
 
 import { combine } from 'lib-common/api'
 import { AttendanceStatus } from 'lib-common/generated/api-types/attendance'
-import { useMutation } from 'lib-common/query'
+import { useMutation, useQueryResult } from 'lib-common/query'
 import { UUID } from 'lib-common/types'
 import useNonNullableParams from 'lib-common/useNonNullableParams'
 import { StaticChip } from 'lib-components/atoms/Chip'
@@ -24,10 +24,11 @@ import { faArrowLeft, faCalendarTimes, faQuestion, farUser } from 'lib-icons'
 import { renderResult } from '../async-rendering'
 import { IconBox } from '../child-attendance/ChildListItem'
 import {
+  childrenQuery,
   deleteChildImageMutation,
   returnToComingMutation
 } from '../child-attendance/queries'
-import { ChildAttendanceContext } from '../child-attendance/state'
+import { useAttendanceStatuses } from '../child-attendance/state'
 import BottomModalMenu from '../common/BottomModalMenu'
 import { FlexColumn } from '../common/components'
 import { useTranslation } from '../common/i18n'
@@ -55,10 +56,8 @@ export default React.memo(function AttendanceChildPage() {
   }>()
 
   const { unitInfoResponse } = useContext(UnitContext)
-
-  const { unitChildren, childAttendanceStatuses } = useContext(
-    ChildAttendanceContext
-  )
+  const unitChildren = useQueryResult(childrenQuery(unitId))
+  const childAttendanceStatuses = useAttendanceStatuses(unitId)
 
   const [uiMode, setUiMode] = useState<
     | 'default'
